feat(auth): refactor authentication handlers and update API constants
