+++ src/modules/auth/basic/commands/handlers/signout.handler.ts
@@ -1,0 +1,12 @@
+import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
+import { SignOutCommand } from '../signout.command';
+import { AuthService } from '../../services/auth.service';
+
+@CommandHandler(SignOutCommand)
+export class SignOutHandler implements ICommandHandler<SignOutCommand> {
+  constructor(private readonly authService: AuthService) {}
+
+  async execute(command: SignOutCommand): Promise<void> {
+    await this.authService.invalidateRefreshToken(command.userId);
+  }
+}
